+++ pkg-manager/resolve-dependencies/src/resolveDependencies.ts
@@ -39,6 +39,7 @@ import {
   type Registries,
   type PkgIdWithPatchHash,
   type PinnedVersion,
+  type PackageVersionPolicy,
 } from '@pnpm/types'
 import * as dp from '@pnpm/dependency-path'
 import { getPreferredVersionsFromLockfileAndManifests } from '@pnpm/lockfile.preferred-versions'
@@ -57,6 +58,8 @@ import { hoistPeers, getHoistableOptionalPeers } from './hoistPeers.js'
 import { wantedDepIsLocallyAvailable } from './wantedDepIsLocallyAvailable.js'
 import { type CatalogLookupMetadata } from './resolveDependencyTree.js'
 import { replaceVersionInBareSpecifier } from './replaceVersionInBareSpecifier.js'
+
+export type { WantedDependency }
 
 const dependencyResolvedLogger = logger('_dependency_resolved')
 
@@ -175,7 +178,7 @@ export interface ResolutionContext {
   missingPeersOfChildrenByPkgId: Record<PkgResolutionId, { depth: number, missingPeersOfChildren: MissingPeersOfChildren }>
   hoistPeers?: boolean
   maximumPublishedBy?: Date
-  minimumReleaseAgeExclude?: (pkgName: string) => boolean
+  publishedByExclude?: PackageVersionPolicy
 }
 
 export interface MissingPeerInfo {
@@ -1303,17 +1306,6 @@ async function resolveDependency (
     if (!options.updateRequested && options.preferredVersion != null) {
       wantedDependency.bareSpecifier = replaceVersionInBareSpecifier(wantedDependency.bareSpecifier, options.preferredVersion)
     }
-    let publishedBy: Date | undefined
-    if (
-      options.publishedBy &&
-      (
-        ctx.minimumReleaseAgeExclude == null ||
-        wantedDependency.alias == null ||
-        !ctx.minimumReleaseAgeExclude(wantedDependency.alias)
-      )
-    ) {
-      publishedBy = options.publishedBy
-    }
     pkgResponse = await ctx.storeController.requestPackage(wantedDependency, {
       alwaysTryWorkspacePackages: ctx.linkWorkspacePackagesDepth >= options.currentDepth,
       currentPkg: currentPkg
@@ -1327,7 +1319,8 @@ async function resolveDependency (
       expectedPkg: currentPkg,
       defaultTag: ctx.defaultTag,
       ignoreScripts: ctx.ignoreScripts,
-      publishedBy,
+      publishedBy: options.publishedBy,
+      publishedByExclude: ctx.publishedByExclude,
       pickLowestVersion: options.pickLowestVersion,
       downloadPriority: -options.currentDepth,
       lockfileDir: ctx.lockfileDir,
@@ -1420,7 +1413,7 @@ async function resolveDependency (
 
   let prepare!: boolean
   let hasBin!: boolean
-  let pkg: PackageManifest = getManifestFromResponse(pkgResponse, wantedDependency)
+  let pkg: PackageManifest = getManifestFromResponse(pkgResponse, wantedDependency, currentPkg)
   if (!pkg.dependencies) {
     pkg.dependencies = {}
   }
@@ -1504,7 +1497,9 @@ async function resolveDependency (
     ) {
       pkg.deprecated = currentPkg.dependencyLockfile.deprecated
     }
-    hasBin = Boolean((pkg.bin && !(pkg.bin === '' || Object.keys(pkg.bin).length === 0)) ?? pkg.directories?.bin)
+    hasBin = (currentPkg.dependencyLockfile?.hasBin != null && !pkg.bin)
+      ? currentPkg.dependencyLockfile.hasBin
+      : Boolean((pkg.bin && !(pkg.bin === '' || Object.keys(pkg.bin).length === 0)) ?? pkg.directories?.bin)
   }
   if (options.currentDepth === 0 && pkgResponse.body.latest && pkgResponse.body.latest !== pkg.version) {
     ctx.outdatedDependencies[pkgResponse.body.id] = pkgResponse.body.latest
@@ -1647,11 +1642,19 @@ async function resolveDependency (
   }
 }
 
-function getManifestFromResponse (
+export function getManifestFromResponse (
   pkgResponse: PackageResponse,
-  wantedDependency: WantedDependency
+  wantedDependency: WantedDependency,
+  currentPkg?: Partial<InfoFromLockfile>
 ): PackageManifest {
   if (pkgResponse.body.manifest) return pkgResponse.body.manifest
+
+  if (currentPkg?.name && currentPkg?.version) {
+    return {
+      name: currentPkg.name,
+      version: currentPkg.version,
+    }
+  }
   return {
     name: wantedDependency.alias ? wantedDependency.alias : wantedDependency.bareSpecifier.split('/').pop()!,
     version: '0.0.0',