+++ env/node.fetcher/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/node.fetcher
 
+## 1001.0.8
+
+### Patch Changes
+
+- @pnpm/fetching.binary-fetcher@1002.0.0
+- @pnpm/tarball-fetcher@1003.0.0
+- @pnpm/node.resolver@1001.0.5
+- @pnpm/create-cafs-store@1000.0.20
+
+## 1001.0.7
+
+### Patch Changes
+
+- @pnpm/node.resolver@1001.0.4
+- @pnpm/tarball-fetcher@1002.0.1
+
 ## 1001.0.6
 
 ### Patch Changes