+++ fetching/fetcher-base/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/fetcher-base
 
+## 1001.0.2
+
+### Patch Changes
+
+- Updated dependencies [7c1382f]
+- Updated dependencies [7c1382f]
+- Updated dependencies [dee39ec]
+  - @pnpm/types@1000.9.0
+  - @pnpm/resolver-base@1005.1.0
+
 ## 1001.0.1
 
 ### Patch Changes