+++ fetching/git-fetcher/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pnpm/git-fetcher
 
+## 1003.0.0
+
+### Patch Changes
+
+- Updated dependencies [8993f68]
+  - @pnpm/worker@1000.3.0
+  - @pnpm/prepare-package@1000.0.26
+  - @pnpm/fetcher-base@1001.0.2
+
+## 1002.0.1
+
+### Patch Changes
+
+- @pnpm/prepare-package@1000.0.25
+
 ## 1002.0.0
 
 ### Patch Changes