+++ packages/core-loggers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/core-loggers
 
+## 1001.0.4
+
+### Patch Changes
+
+- Updated dependencies [7c1382f]
+- Updated dependencies [dee39ec]
+  - @pnpm/types@1000.9.0
+
 ## 1001.0.3
 
 ### Patch Changes