+++ packages/dependency-path/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/dependency-path
 
+## 1001.1.3
+
+### Patch Changes
+
+- Updated dependencies [7c1382f]
+- Updated dependencies [dee39ec]
+  - @pnpm/types@1000.9.0
+  - @pnpm/crypto.hash@1000.2.1
+
 ## 1001.1.2
 
 ### Patch Changes