+++ reviewing/plugin-commands-listing/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pnpm/plugin-commands-listing
 
+## 1000.1.7
+
+### Patch Changes
+
+- Updated dependencies [7c1382f]
+- Updated dependencies [dee39ec]
+  - @pnpm/types@1000.9.0
+  - @pnpm/cli-utils@1001.2.7
+  - @pnpm/config@1004.4.2
+  - @pnpm/list@1000.1.3
+
+## 1000.1.6
+
+### Patch Changes
+
+- Updated dependencies [9865167]
+  - @pnpm/config@1004.4.1
+  - @pnpm/cli-utils@1001.2.6
+
 ## 1000.1.5
 
 ### Patch Changes