+++ testing/temp-store/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/testing.temp-store
 
+## 1000.0.22
+
+### Patch Changes
+
+- Updated dependencies [7c1382f]
+  - @pnpm/store-controller-types@1004.1.0
+  - @pnpm/package-store@1004.0.0
+  - @pnpm/client@1001.1.3
+
+## 1000.0.21
+
+### Patch Changes
+
+- @pnpm/client@1001.1.2
+- @pnpm/package-store@1003.0.0
+
 ## 1000.0.20
 
 ### Patch Changes